feat(mllm-v2): Support single-controller for Megatron-Inference. (#4009)

+++ tests/unit/data/datasets/test_mmpr_tiny.py
@@ -307,7 +307,7 @@ class TestVLMProcessorMMPRTiny:
             torch.tensor([[224, 224], [224, 224], [224, 224]]),
         )
         assert torch.equal(
-            user_message["num_frames"].as_tensor(), torch.ones(3, dtype=torch.long)
+            user_message["num_frames"].as_tensor(), torch.ones(3, dtype=torch.int32)
         )
 
     def test_prompted_text_contains_boxed_literal_and_no_raw_dataset_string(
